+++ resources/webpack.config.js
@@ -21,31 +21,19 @@ module.exports = {
         test: /\.(txt|html)/,
         type: "asset/source",
       },
-      {
-        // asset/source exports the source code of the asset.
-        resourceQuery: /staticText/,
-        type: "asset/source",
-      },
-      {
-        // asset/inline exports the raw bytes of the asset.
-        // We base64 encode them here
-        resourceQuery: /staticBinary/,
-        type: "asset/inline",
-        generator: {
-          /**
-           * @param {Buffer} content
-           * @returns {string}
-           */
-          dataUrl: content => {
-            return content.toString('base64');
-          },
-        }
-      },
     ],
   },
   plugins: [
     // Polyfills go here.
     // Used for, e.g., any cross-platform WHATWG,
     // or core nodejs modules needed for your application.
+  ],
+  externals: [
+    ({request,}, callback) => {
+      if (/^fastly:.*$/.test(request)) {
+        return callback(null, 'commonjs ' + request);
+      }
+      callback();
+    }
   ],
 };